Port typed VolumeDataRequest templates from HueSpace.

+++ java/src/org/opengroup/openvds/VolumeDataAccessManager.java
@@ -2335,7 +2335,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
     
     
 
-    native private long RequestProjectedVolumeSubsetByteImpl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, int interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
+    native private long RequestProjectedVolumeSubsetByteImpl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, long interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
 
     /**
      * Request a subset projected from an arbitrary 3D plane through the subset onto one of the sides of the subset.
@@ -2359,7 +2359,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
     }
     
 
-    native private long RequestProjectedVolumeSubsetByte2Impl(long native_object, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, int interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
+    native private long RequestProjectedVolumeSubsetByte2Impl(long native_object, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, long interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
 
     /**
      * Request a subset projected from an arbitrary 3D plane through the subset onto one of the sides of the subset, using an automatically allocated buffer.
@@ -2382,7 +2382,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
     }
     
 
-    native private long RequestProjectedVolumeSubsetUShortImpl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, int interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
+    native private long RequestProjectedVolumeSubsetUShortImpl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, long interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
 
     /**
      * Request a subset projected from an arbitrary 3D plane through the subset onto one of the sides of the subset.
@@ -2406,7 +2406,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
     }
     
 
-    native private long RequestProjectedVolumeSubsetUShort2Impl(long native_object, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, int interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
+    native private long RequestProjectedVolumeSubsetUShort2Impl(long native_object, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, long interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
 
     /**
      * Request a subset projected from an arbitrary 3D plane through the subset onto one of the sides of the subset, using an automatically allocated buffer.
@@ -2429,7 +2429,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
     }
     
 
-    native private long RequestProjectedVolumeSubsetUIntImpl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, int interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
+    native private long RequestProjectedVolumeSubsetUIntImpl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, long interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
 
     /**
      * Request a subset projected from an arbitrary 3D plane through the subset onto one of the sides of the subset.
@@ -2453,7 +2453,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
     }
     
 
-    native private long RequestProjectedVolumeSubsetUInt2Impl(long native_object, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, int interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
+    native private long RequestProjectedVolumeSubsetUInt2Impl(long native_object, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, long interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
 
     /**
      * Request a subset projected from an arbitrary 3D plane through the subset onto one of the sides of the subset, using an automatically allocated buffer.
@@ -2476,7 +2476,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
     }
     
 
-    native private long RequestProjectedVolumeSubsetULongImpl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, int interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
+    native private long RequestProjectedVolumeSubsetULongImpl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, long interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
 
     /**
      * Request a subset projected from an arbitrary 3D plane through the subset onto one of the sides of the subset.
@@ -2500,7 +2500,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
     }
     
 
-    native private long RequestProjectedVolumeSubsetULong2Impl(long native_object, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, int interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
+    native private long RequestProjectedVolumeSubsetULong2Impl(long native_object, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, long interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
 
     /**
      * Request a subset projected from an arbitrary 3D plane through the subset onto one of the sides of the subset, using an automatically allocated buffer.
@@ -2523,7 +2523,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
     }
     
 
-    native private long RequestProjectedVolumeSubsetFloatImpl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, int interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
+    native private long RequestProjectedVolumeSubsetFloatImpl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, long interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
 
     /**
      * Request a subset projected from an arbitrary 3D plane through the subset onto one of the sides of the subset.
@@ -2547,7 +2547,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
     }
     
 
-    native private long RequestProjectedVolumeSubsetFloat2Impl(long native_object, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, int interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
+    native private long RequestProjectedVolumeSubsetFloat2Impl(long native_object, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, long interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
 
     /**
      * Request a subset projected from an arbitrary 3D plane through the subset onto one of the sides of the subset, using an automatically allocated buffer.
@@ -2570,7 +2570,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
     }
     
 
-    native private long RequestProjectedVolumeSubsetDoubleImpl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, int interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
+    native private long RequestProjectedVolumeSubsetDoubleImpl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, long interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
 
     /**
      * Request a subset projected from an arbitrary 3D plane through the subset onto one of the sides of the subset.
@@ -2594,7 +2594,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
     }
     
 
-    native private long RequestProjectedVolumeSubsetDouble2Impl(long native_object, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, int interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
+    native private long RequestProjectedVolumeSubsetDouble2Impl(long native_object, long dimensionsND, int LOD, int channel, int[] minVoxelCoordinates, int[] maxVoxelCoordinates, ByteBuffer voxelPlane, long voxelPlane_byteoffset, long projectedDimensions, long interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
 
     /**
      * Request a subset projected from an arbitrary 3D plane through the subset onto one of the sides of the subset, using an automatically allocated buffer.
@@ -2619,7 +2619,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
 
 
     ///AUTOGEN-IGNORE: CXX_METHOD RequestVolumeSamples std::shared_ptr<OpenVDS::VolumeDataRequest_t<float>> (float *, int64_t, OpenVDS::DimensionsND, int, int, float const (*)[6], int, OpenVDS::InterpolationMethod, OpenVDS::optional<float>) FUNCTIONPROTO
-    native private long RequestVolumeSamplesImpl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, ByteBuffer samplePositions, int interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
+    native private long RequestVolumeSamplesImpl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, ByteBuffer samplePositions, long interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
 
     /**
      * Request sampling of the input VDS at the specified coordinates.
@@ -2639,7 +2639,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
 
 
     ///AUTOGEN-IGNORE: CXX_METHOD RequestVolumeSamples std::shared_ptr<OpenVDS::VolumeDataRequest_t<float>> (OpenVDS::DimensionsND, int, int, float const (*)[6], int, OpenVDS::InterpolationMethod, OpenVDS::optional<float>) FUNCTIONPROTO
-    native private long RequestVolumeSamples2Impl(long native_object, long dimensionsND, int LOD, int channel, ByteBuffer samplePositions, int interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
+    native private long RequestVolumeSamples2Impl(long native_object, long dimensionsND, int LOD, int channel, ByteBuffer samplePositions, long interpolationMethod, float replacementNoValue, boolean use_replacementNoValue);
 
     /**
      * Request sampling of the input VDS at the specified coordinates, using an automatically allocated buffer.
@@ -2659,7 +2659,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
     
     
     ///AUTOGEN-IGNORE: CXX_METHOD RequestVolumeTraces std::shared_ptr<OpenVDS::VolumeDataRequest_t<float>> (OpenVDS::DimensionsND, int, int, float const (*)[6], int, OpenVDS::InterpolationMethod, int, OpenVDS::optional<float>) FUNCTIONPROTO
-    native private long RequestVolumeTracesImpl(long native_object, long dimensionsND, int LOD, int channel, ByteBuffer tracePositions, int interpolationMethod, int traceDimension, float replacementNoValue, boolean use_replacementNoValue);
+    native private long RequestVolumeTracesImpl(long native_object, long dimensionsND, int LOD, int channel, ByteBuffer tracePositions, long interpolationMethod, int traceDimension, float replacementNoValue, boolean use_replacementNoValue);
 
     /**
      * Request traces from the input VDS, using an automatically allocated buffer.
@@ -2679,7 +2679,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
 
 
     ///AUTOGEN-IGNORE: CXX_METHOD RequestVolumeTraces std::shared_ptr<OpenVDS::VolumeDataRequest_t<float>> (float *, int64_t, OpenVDS::DimensionsND, int, int, float const (*)[6], int, OpenVDS::InterpolationMethod, int, OpenVDS::optional<float>) FUNCTIONPROTO
-    native private long RequestVolumeTraces2Impl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, ByteBuffer tracePositions, int interpolationMethod, int traceDimension, float replacementNoValue, boolean use_replacementNoValue);
+    native private long RequestVolumeTraces2Impl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, ByteBuffer tracePositions, long interpolationMethod, int traceDimension, float replacementNoValue, boolean use_replacementNoValue);
 
     /**
      * Request traces from the input VDS.
@@ -2703,7 +2703,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
 
 // FIXME? RequestVolumeTraceRanges does not exist in OpenVDS 
 //    ///AUTOGEN-IGNORE: CXX_METHOD RequestVolumeTraceRanges std::shared_ptr<OpenVDS::VolumeDataRequest_t<float>> (OpenVDS::DimensionsND, int, int, float const (*)[6], int, OpenVDS::InterpolationMethod, int, int, int, OpenVDS::optional<float>) FUNCTIONPROTO
-//    native private long RequestVolumeTraceRangesImpl(long native_object, long dimensionsND, int LOD, int channel, ByteBuffer tracePositions, int traceCount, int interpolationMethod, int traceDimension, int traceMin, int traceMax, float replacementNoValue, boolean use_replacementNoValue);
+//    native private long RequestVolumeTraceRangesImpl(long native_object, long dimensionsND, int LOD, int channel, ByteBuffer tracePositions, int traceCount, long interpolationMethod, int traceDimension, int traceMin, int traceMax, float replacementNoValue, boolean use_replacementNoValue);
 //
 //    /**
 //     * Request traces from the input VDS in a given range of samples, using an automatically allocated buffer.
@@ -2727,7 +2727,7 @@ of the number of chunks in some of the dimensions. Do not change this from the d
 //    ///FIXME AUTOGEN-ADD-OVERLOAD: VolumeDataRequestFloat requestVolumeTraceRanges(DimensionsND dimensionsND, int LOD, int channel, NDPosArray tracePositions, int traceCount, InterpolationMethod interpolationMethod, int traceDimension, int traceMin, int traceMax, Float replacementNoValue) -> replacementNoValue=null
 //
 //    ///AUTOGEN-IGNORE: CXX_METHOD RequestVolumeTraceRanges std::shared_ptr<OpenVDS::VolumeDataRequest_t<float>> (float *, int64_t, OpenVDS::DimensionsND, int, int, float const (*)[6], int, OpenVDS::InterpolationMethod, int, int, int, OpenVDS::optional<float>) FUNCTIONPROTO
-//    native private long RequestVolumeTraceRanges2Impl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, ByteBuffer tracePositions, int traceCount, int interpolationMethod, int traceDimension, int traceMin, int traceMax, float replacementNoValue, boolean use_replacementNoValue);
+//    native private long RequestVolumeTraceRanges2Impl(long native_object, ByteBuffer buffer, long dimensionsND, int LOD, int channel, ByteBuffer tracePositions, int traceCount, long interpolationMethod, int traceDimension, int traceMin, int traceMax, float replacementNoValue, boolean use_replacementNoValue);
 //
 //    /**
 //     * Request traces from the input VDS in a given range of samples.
